feat: #2661 moved default data to seperate files
