+++ apps/api/src/app/candidate-documents/default-candidate-documents.ts
@@ -1,0 +1,16 @@
+import { ICandidateDocument } from '@gauzy/models';
+
+export const DEFAULT_CANDIDATE_DOCUMENTS: ICandidateDocument[] = [
+	{
+		name: 'Document 1',
+		documentUrl:
+			'http://res.cloudinary.com/evereq/image/upload/v1587742725/everbie-products-images/n07vjqa8pa8dfinkzqdy.pdf',
+		tenant: {}
+	},
+	{
+		name: 'Document 2',
+		documentUrl:
+			'http://res.cloudinary.com/evereq/raw/upload/v1587742757/everbie-products-images/wxjghcvuysc3imrx7z2t.docx',
+		tenant: {}
+	}
+];
