+++ apps/api/src/app/goal-kpi-template/default-goal-kpi-templates.ts
@@ -1,0 +1,22 @@
+import { IGoalKPITemplate } from '@gauzy/models';
+
+export const DEFAULT_GOAL_KPI_TEMPLATES: IGoalKPITemplate[] = [
+	{
+		name: 'Average response time',
+		description: '',
+		type: 'Numerical',
+		unit: 'ms',
+		operator: '<=',
+		currentValue: 1000,
+		targetValue: 500
+	},
+	{
+		name: '# of Priority bugs in production',
+		description: '',
+		type: 'Numerical',
+		unit: 'bugs',
+		operator: '<=',
+		currentValue: 15,
+		targetValue: 2
+	}
+];
