+++ apps/api/src/app/goal-kpi/default-goal-kpis.ts
@@ -1,0 +1,22 @@
+import { IKPI } from '@gauzy/models';
+
+export const DEFAULT_GOAL_KPIS: IKPI[] = [
+	{
+		name: 'Average response time',
+		description: '',
+		type: 'Numerical',
+		unit: 'ms',
+		operator: '<=',
+		currentValue: 1000,
+		targetValue: 500
+	},
+	{
+		name: '# of Priority bugs in production',
+		description: '',
+		type: 'Numerical',
+		unit: 'bugs',
+		operator: '<=',
+		currentValue: 15,
+		targetValue: 2
+	}
+];
