+++ apps/api/src/app/goal-kpi/goal-kpi.seed.ts
@@ -3,27 +3,7 @@ import { Tenant } from '../tenant/tenant.entity';
 import { IEmployee, IOrganization } from '@gauzy/models';
 import { GoalKPI } from './goal-kpi.entity';
 import * as faker from 'faker';
-
-const goalKPIData = [
-	{
-		name: 'Average response time',
-		description: '',
-		type: 'Numerical',
-		unit: 'ms',
-		operator: '<=',
-		currentValue: 1000,
-		targetValue: 500
-	},
-	{
-		name: '# of Priority bugs in production',
-		description: '',
-		type: 'Numerical',
-		unit: 'bugs',
-		operator: '<=',
-		currentValue: 15,
-		targetValue: 2
-	}
-];
+import { DEFAULT_GOAL_KPIS } from './default-goal-kpis';
 
 export const createDefaultGoalKpi = async (
 	connection: Connection,
@@ -32,36 +12,28 @@ export const createDefaultGoalKpi = async (
 	employees: IEmployee[]
 ): Promise<GoalKPI[]> => {
 	const goalKpis: GoalKPI[] = [];
-
 	organizations.forEach((organization: IOrganization) => {
-		goalKPIData.forEach((goalKPI) => {
-			const goalkpi = new GoalKPI();
-			goalkpi.name = goalKPI.name;
-			goalkpi.description = ' ';
-			goalkpi.type = goalKPI.type;
-			goalkpi.operator = goalKPI.operator;
-			goalkpi.unit = goalKPI.unit;
-			goalkpi.lead = faker.random.arrayElement(employees);
-			goalkpi.currentValue = goalKPI.currentValue;
-			goalkpi.targetValue = goalKPI.targetValue;
-			goalkpi.organization = organization;
-			goalkpi.tenant = tenant;
-			goalKpis.push(goalkpi);
+		DEFAULT_GOAL_KPIS.forEach((goalKPI) => {
+			const goalKpi = new GoalKPI();
+			goalKpi.name = goalKPI.name;
+			goalKpi.description = ' ';
+			goalKpi.type = goalKPI.type;
+			goalKpi.operator = goalKPI.operator;
+			goalKpi.unit = goalKPI.unit;
+			goalKpi.lead = faker.random.arrayElement(employees);
+			goalKpi.currentValue = goalKPI.currentValue;
+			goalKpi.targetValue = goalKPI.targetValue;
+			goalKpi.organization = organization;
+			goalKpi.tenant = tenant;
+			goalKpis.push(goalKpi);
 		});
 	});
-
-	await insertRandomGoalKpi(connection, goalKpis);
-	return goalKpis;
+	return await insertRandomGoalKpi(connection, goalKpis);
 };
 
 const insertRandomGoalKpi = async (
 	connection: Connection,
 	goalKpis: GoalKPI[]
-) => {
-	await connection
-		.createQueryBuilder()
-		.insert()
-		.into(GoalKPI)
-		.values(goalKpis)
-		.execute();
+): Promise<GoalKPI[]> => {
+	return await connection.manager.save(goalKpis);
 };
