+++ apps/api/src/app/goal-template/goal-template.seed.ts
@@ -1,36 +1,8 @@
 import { Connection } from 'typeorm';
 import { GoalTemplate } from './goal-template.entity';
-import { GoalTemplateCategoriesEnum } from '@gauzy/models';
 import { Organization } from '../organization/organization.entity';
 import { Tenant } from '../tenant/tenant.entity';
-
-const defaultGoalTemplateData = [
-	{
-		name: 'Improve product performance',
-		level: 'Organization',
-		category: GoalTemplateCategoriesEnum.PRODUCT_MANAGEMENT
-	},
-	{
-		name: 'Successfully launch version 2 of our main product',
-		level: 'Organization',
-		category: GoalTemplateCategoriesEnum.MARKETING
-	},
-	{
-		name: 'Redesign and launch our new landing page',
-		level: 'Team',
-		category: GoalTemplateCategoriesEnum.PRODUCT_MANAGEMENT
-	},
-	{
-		name: 'Increase quality of releases and make sure they are timely',
-		level: 'Team',
-		category: GoalTemplateCategoriesEnum.PRODUCT_MANAGEMENT
-	},
-	{
-		name: 'Identify problems with current user interface',
-		level: 'Employee',
-		category: GoalTemplateCategoriesEnum.PRODUCT_MANAGEMENT
-	}
-];
+import { DEFAULT_GOAL_TEMPLATES } from './default-goal-templates';
 
 export const createDefaultGoalTemplates = async (
 	connection: Connection,
@@ -38,7 +10,7 @@ export const createDefaultGoalTemplates = async (
 	organization: Organization
 ): Promise<GoalTemplate[]> => {
 	const defaultGoalTemplates = [];
-	defaultGoalTemplateData.forEach((goalData) => {
+	DEFAULT_GOAL_TEMPLATES.forEach((goalData) => {
 		const goalTemplate = new GoalTemplate();
 		goalTemplate.name = goalData.name;
 		goalTemplate.level = goalData.level;
