+++ apps/api/src/app/goal/default-goals.ts
@@ -1,0 +1,22 @@
+export const DEFAULT_GOALS = [
+	{
+		name: 'Improve product performance',
+		level: 'Organization'
+	},
+	{
+		name: 'Successfully launch version 2 of our main product',
+		level: 'Organization'
+	},
+	{
+		name: 'Redesign and launch our new landing page',
+		level: 'Team'
+	},
+	{
+		name: 'Increase quality of releases and make sure they are timely',
+		level: 'Team'
+	},
+	{
+		name: 'Identify problems with current user interface',
+		level: 'Employee'
+	}
+];
