+++ apps/api/src/app/help-center/help-center.seed.ts
@@ -3,112 +3,7 @@ import { IHelpCenter } from '@gauzy/models';
 import { Connection } from 'typeorm';
 import { Tenant } from '../tenant/tenant.entity';
 import { Organization } from '../organization/organization.entity';
-
-const helpCenterMenuList: IHelpCenter[] = [
-	{
-		name: 'Gauzy Platform',
-		icon: 'book-open-outline',
-		flag: 'base',
-		privacy: 'eye-outline',
-		language: 'en',
-		color: '#d53636',
-		index: 0,
-		tenant: {},
-		children: [
-			{
-				name: 'Cookies',
-				icon: 'alert-circle-outline',
-				flag: 'category',
-				privacy: 'eye-outline',
-				description: 'Information',
-				language: 'en',
-				color: '#d53636',
-				tenant: {},
-				index: 0
-			},
-			{
-				name: 'Device',
-				icon: 'book-open-outline',
-				flag: 'category',
-				privacy: 'eye-off-outline',
-				description: 'Device Information',
-				language: 'en',
-				color: '#d53636',
-				tenant: {},
-				index: 1
-			},
-			{
-				flag: 'category',
-				icon: 'book-open-outline',
-				privacy: 'eye-off-outline',
-				name: 'Privacy',
-				description: 'Gauzy Privacy Statement',
-				data: 'Usage Information',
-				language: 'en',
-				color: '#d53636',
-				tenant: {},
-				index: 2
-			},
-			{
-				flag: 'category',
-				icon: 'book-open-outline',
-				privacy: 'eye-off-outline',
-				name: 'Testing',
-				description: 'Gauzy Testing',
-				language: 'en',
-				color: '#d53636',
-				tenant: {},
-				index: 3
-			}
-		]
-	},
-	{
-		name: 'Ever Platform',
-		icon: 'book-open-outline',
-		flag: 'base',
-		privacy: 'eye-off-outline',
-		language: 'en',
-		color: '#d53636',
-		index: 1,
-		tenant: {},
-		children: [
-			{
-				name: 'Cookies',
-				icon: 'alert-circle-outline',
-				flag: 'category',
-				privacy: 'eye-outline',
-				description: 'Information',
-				language: 'en',
-				color: '#d53636',
-				tenant: {},
-				index: 0
-			}
-		]
-	},
-	{
-		flag: 'base',
-		icon: 'book-open-outline',
-		privacy: 'eye-off-outline',
-		name: 'Privacy',
-		language: 'en',
-		color: '#d53636',
-		index: 2,
-		tenant: {},
-		children: [
-			{
-				name: 'Cookies',
-				icon: 'alert-circle-outline',
-				flag: 'category',
-				privacy: 'eye-outline',
-				description: 'Information',
-				language: 'en',
-				color: '#d53636',
-				tenant: {},
-				index: 0
-			}
-		]
-	}
-];
+import { DEFAULT_HELP_CENTER_MENUS } from './default-help-centers';
 
 export const createHelpCenter = async (
 	connection: Connection,
@@ -120,6 +15,7 @@ export const createHelpCenter = async (
 		org: Organization;
 	}
 ): Promise<IHelpCenter[]> => {
+	const helpCenterMenuList: IHelpCenter[] = DEFAULT_HELP_CENTER_MENUS;
 	for (const node of helpCenterMenuList) {
 		const helpCenter: HelpCenter = {
 			...node,
@@ -134,7 +30,6 @@ export const createHelpCenter = async (
 		const entity = await createEntity(connection, helpCenter);
 		await save(connection, entity);
 	}
-
 	return helpCenterMenuList;
 };
