+++ apps/api/src/app/organization-employment-type/organization-employment-type.seed.ts
@@ -4,6 +4,7 @@ import { OrganizationEmploymentType } from './organization-employment-type.entit
 import { Organization } from '../organization/organization.entity';
 import { Employee } from '../employee/employee.entity';
 import { Tenant } from '../tenant/tenant.entity';
+import { DEFAULT_ORGANIZATION_TEAMS } from '../organization-team/default-organization-teams';
 
 export const seedDefaultEmploymentTypes = async (
 	connection: Connection,
@@ -11,45 +12,7 @@ export const seedDefaultEmploymentTypes = async (
 	employees: Employee[],
 	defaultOrganization: Organization
 ) => {
-	const defaultTeams = [
-		{
-			name: 'Employees',
-			defaultMembers: [
-				'admin@ever.co',
-				'ruslan@ever.co',
-				'alish@ever.co',
-				'blagovest@ever.co',
-				'elvis@ever.co',
-				'hristo@ever.co',
-				'alex@ever.co',
-				'pavel@ever.co',
-				'yavor@ever.co',
-				'tsvetelina@ever.co',
-				'julia@ever.co'
-			],
-			manager: ['ruslan@ever.co']
-		},
-		{
-			name: 'Contractors',
-			defaultMembers: [
-				'dimana@ever.co',
-				'deko898@hotmail.com',
-				'muiz@smooper.xyz',
-				'ckhandla94@gmail.com'
-			],
-			manager: ['ruslan@ever.co', 'rachit@ever.co']
-		},
-		{
-			name: 'Designers',
-			defaultMembers: ['julia@ever.co', 'yordan@ever.co'],
-			manager: []
-		},
-		{
-			name: 'QA',
-			defaultMembers: ['julia@ever.co', 'yordan@ever.co'],
-			manager: []
-		}
-	];
+	const defaultTeams = DEFAULT_ORGANIZATION_TEAMS;
 	const fullTimeEmployees = defaultTeams[0].defaultMembers;
 	const contractors = defaultTeams[1].defaultMembers;
 	const employmentTypes: OrganizationEmploymentType[] = Object.values(
