+++ apps/api/src/app/organization-team/default-organization-teams.ts
@@ -1,0 +1,39 @@
+export const DEFAULT_ORGANIZATION_TEAMS = [
+	{
+		name: 'Employees',
+		defaultMembers: [
+			'admin@ever.co',
+			'ruslan@ever.co',
+			'alish@ever.co',
+			'blagovest@ever.co',
+			'elvis@ever.co',
+			'hristo@ever.co',
+			'alex@ever.co',
+			'pavel@ever.co',
+			'yavor@ever.co',
+			'tsvetelina@ever.co',
+			'julia@ever.co'
+		],
+		manager: ['ruslan@ever.co']
+	},
+	{
+		name: 'Contractors',
+		defaultMembers: [
+			'dimana@ever.co',
+			'deko898@hotmail.com',
+			'muiz@smooper.xyz',
+			'ckhandla94@gmail.com'
+		],
+		manager: ['ruslan@ever.co', 'rachit@ever.co']
+	},
+	{
+		name: 'Designers',
+		defaultMembers: ['julia@ever.co', 'yordan@ever.co'],
+		manager: []
+	},
+	{
+		name: 'QA',
+		defaultMembers: ['julia@ever.co', 'yordan@ever.co'],
+		manager: []
+	}
+];
