+++ apps/api/src/app/organization_employeeLevel/default-organization-employee-levels.ts
@@ -1,0 +1,14 @@
+export const DEFAULT_EMPLOYEE_LEVELS = [
+	{
+		level: 'Level A'
+	},
+	{
+		level: 'Level B'
+	},
+	{
+		level: 'Level C'
+	},
+	{
+		level: 'Level D'
+	}
+];
